refactor(checkpoint): move grouped params + builders into their own module
The three builders added alongside CaptureFsQuickParams, ForkParams and
CaptureVmFullParams pushed checkpoint/mod.rs to 1674 production lines,
past the 1500-line gate. Roughly four hundred of those lines are setters
that say nothing about how a checkpoint is taken, so they move to
checkpoint/params.rs with their tests and are re-exported unchanged.
